@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0  <3.0.0"
+  sdk: ">=2.14.0  <3.0.0"
 
 dependencies:
   cupertino_icons: ^1.0.2
@@ -24,6 +24,8 @@ dependencies:
   cached_network_image: ^3.1.0
   auto_size_text: ^3.0.0-nullsafety.0
   logger: 
+  url_launcher: ^6.0.12
+  flutter_dotenv: ^5.0.2
 
 
 dev_dependencies:
@@ -33,6 +35,8 @@ dev_dependencies:
   stacked_generator:
 
 flutter:
+  assets:
+    - .env
 
   uses-material-design: true
 
