@@ -4,6 +4,7 @@ import 'package:designsys/designsys.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:zawadi_design/models/article_model.dart';
 import 'package:zawadi_design/pages/news/widgets/article_view.dart';
+import 'package:zawadi_design/ui/web_assets.dart';
 
 ///This is the widget in front page displaying a slighlty larger card of news
 ///currently only a single widgets is displayed in mobile view
@@ -11,7 +12,7 @@ class ArticleCardLarge extends StatelessWidget {
   const ArticleCardLarge(this.article, this.source);
 
   @required
-  final Article article;
+  final Article? article;
 
   final bool source;
 
@@ -19,7 +20,7 @@ class ArticleCardLarge extends StatelessWidget {
   Widget build(BuildContext context) {
     return GestureDetector(
       onTap: () =>
-          ZawadiMethods.viewTransition(context, ArticleView(article, source)),
+          ZawadiMethods.viewTransition(context, ArticleView(article!, source)),
       child: Card(
         color: Colors.brown[200],
         margin: EdgeInsets.all(8),
@@ -31,7 +32,7 @@ class ArticleCardLarge extends StatelessWidget {
             ClipSquircleBorder(
               radius: BorderRadius.all(Radius.circular(25)),
               child: CachedNetworkImage(
-                imageUrl: article.getImageUrl,
+                imageUrl: article?.getImageUrl ?? WebAssets().dummyImage,
                 fit: BoxFit.cover,
                 height: 200,
                 placeholder: (context, url) => Loading(
@@ -39,16 +40,11 @@ class ArticleCardLarge extends StatelessWidget {
                 ),
                 errorWidget: (context, url, error) => Icon(Icons.error),
               ),
-              // Image.network(
-              //   article.getImageUrl,
-              //   fit: BoxFit.cover,
-              //   height: 200,
-              // ),
             ),
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 5, horizontal: 12),
               child: AutoSizeText(
-                article.name!,
+                article?.name ?? "",
                 style: GoogleFonts.notoSans(
                   textStyle: Theme.of(context).textTheme.bodyText2!.copyWith(
                         color: Theme.of(context).primaryColor,
@@ -62,8 +58,9 @@ class ArticleCardLarge extends StatelessWidget {
               padding: const EdgeInsets.symmetric(horizontal: 12),
               child: Text(
                 !source
-                    ? article.title!.substring(0, article.title!.indexOf(' - '))
-                    : article.title!,
+                    ? article?.title ??
+                        "".substring(0, article?.title!.indexOf(' - '))
+                    : article?.title ?? "",
                 style: GoogleFonts.notoSans(
                   fontSize: 19,
                   fontWeight: FontWeight.w600,
@@ -74,7 +71,7 @@ class ArticleCardLarge extends StatelessWidget {
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 12),
               child: Text(
-                article.getPublishedAtFormattedTime,
+                article?.getPublishedAtFormattedTime ?? '00:00:00:00',
                 style: Theme.of(context).textTheme.caption!.copyWith(
                       color: Theme.of(context).primaryColor,
                       fontWeight: FontWeight.w400,
