@@ -4,6 +4,7 @@ import 'package:designsys/designsys.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:zawadi_design/models/article_model.dart';
 import 'package:zawadi_design/pages/news/widgets/article_view.dart';
+import 'package:zawadi_design/ui/web_assets.dart';
 
 ///This widget is medeum sized in shape and follows after larger card in
 ///front page
@@ -11,7 +12,7 @@ class ArticleCardMedium extends StatelessWidget {
   const ArticleCardMedium(this.article, this.source);
 
   @required
-  final Article article;
+  final Article? article;
 
   final bool source;
 
@@ -37,7 +38,7 @@ class ArticleCardMedium extends StatelessWidget {
                     mainAxisAlignment: MainAxisAlignment.start,
                     children: [
                       AutoSizeText(
-                        article.name!,
+                        article?.name ?? '',
                         style: GoogleFonts.notoSans(
                           textStyle:
                               Theme.of(context).textTheme.bodyText2!.copyWith(
@@ -52,9 +53,10 @@ class ArticleCardMedium extends StatelessWidget {
                         width: MediaQuery.of(context).size.width - 64,
                         child: AutoSizeText(
                           !source
-                              ? article.title!
-                                  .substring(0, article.title!.indexOf(' - '))
-                              : article.title!,
+                              ? article?.title ??
+                                  ''.substring(
+                                      0, article?.title!.indexOf(' - '))
+                              : article?.title ?? '',
                           style: GoogleFonts.notoSans(
                             // fontSize: 19,
                             fontWeight: FontWeight.w800,
@@ -69,7 +71,7 @@ class ArticleCardMedium extends StatelessWidget {
                       Container(
                         width: MediaQuery.of(context).size.width - 164,
                         child: Text(
-                          article.getPublishedAtFormattedTime,
+                          article?.getPublishedAtFormattedTime ?? '',
                           style: Theme.of(context).textTheme.caption!.copyWith(
                                 color: Theme.of(context).primaryColor,
                                 fontWeight: FontWeight.w400,
@@ -85,7 +87,7 @@ class ArticleCardMedium extends StatelessWidget {
                 child: ClipSquircleBorder(
                   radius: BorderRadius.all(Radius.circular(30)),
                   child: CachedNetworkImage(
-                    imageUrl: article.getImageUrl,
+                    imageUrl: article?.getImageUrl ?? WebAssets().dummyImage,
                     fit: BoxFit.cover,
                     height: 100,
                     placeholder: (context, url) => Loading(
