@@ -206,10 +206,10 @@ class _ZawadiExpansionTileState extends State<ZawadiExpansionTile>
     _borderColorTween..end = theme.dividerColor;
     _headerColorTween
       ..begin = theme.textTheme.subtitle1!.color
-      ..end = theme.accentColor;
+      ..end = theme.colorScheme.secondary;
     _iconColorTween
       ..begin = theme.unselectedWidgetColor
-      ..end = theme.accentColor;
+      ..end = theme.colorScheme.secondary;
     _backgroundColorTween..end = widget.backgroundColor;
     super.didChangeDependencies();
   }
