@@ -1,7 +1,6 @@
 import 'package:designsys/designsys.dart';
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:stacked_services/stacked_services.dart';
 import 'package:zawadi_design/app/app.locator.dart';
 import 'package:zawadi_design/app/app.router.dart';
@@ -24,21 +23,12 @@ Future<void> main() async {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    SystemChrome.setEnabledSystemUIMode(SystemUiMode.edgeToEdge);
-    SystemChrome.setSystemUIOverlayStyle(
-      SystemUiOverlayStyle(
-        statusBarColor: Colors.transparent,
-        statusBarBrightness: Brightness.dark,
-        statusBarIconBrightness: Brightness.dark,
-        systemNavigationBarColor: Colors.transparent,
-        systemNavigationBarIconBrightness: Brightness.light,
-      ),
-    );
     return ZawadiApp(
       title: "",
       primaryLightColor: Colors.teal,
       primaryDarkColor: Colors.teal,
       builder: DevicePreview.appBuilder, // Add the builder here
+
       navigatorKey: StackedService.navigatorKey,
       onGenerateRoute: StackedRouter().onGenerateRoute,
     );
